+++ delve/tests/to_str.rs
@@ -23,13 +23,13 @@ fn test_attrs() {
     #[derive(Debug, PartialEq, EnumToStr)]
     #[allow(dead_code)]
     #[delve(rename_variants = "uppercase")]
-    enum Week {
+    enum Week<'a> {
         Sunday,
         Monday,
         Tuesday,
         #[delve(skip)]
         Wednesday,
-        Thursday,
+        Thursday(&'a String),
         #[delve(to = "fri")]
         Friday,
         Saturday,
@@ -37,6 +37,8 @@ fn test_attrs() {
 
     assert_eq!("MONDAY", <Week as Into<&'static str>>::into(Week::Monday));
     assert_eq!("fri", <Week as Into<&'static str>>::into(Week::Friday));
+
+    assert_eq!("fri", <&Week as Into<&'static str>>::into(&Week::Friday));
 }
 
 #[test]