+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0 (`delve-derive`)
+
+- Fixed `EnumToStr` to also implement `From<&TheEnum> to &'static str`.
+
 ## 0.2.0 (`delve-derive`)
 
 - Added `resolve` to container attributes to fix multiple type aliases poiting to the same type being considered different types.